requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Security & Authentication
argon2-cffi==23.1.0
pyotp==2.9.0
cryptography==41.0.7

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Broker SDKs
# smartapi-python==1.3.0  # AngelOne
# kiteconnect==4.2.0      # Zerodha
# dhanhq==1.3.3           # Dhan

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
